+++ docs/changelog.md
@@ -6,6 +6,7 @@
 - Adds the ability to include/exclude relationships by the tag of source of destination elements (e.g. `exclude "* -> element.tag==Tag"`).
 - Adds support for directory based includes with `!include`.
 - Adds support for element expressions to be used in the afferent/efferent coupling expression (e.g. `exclude "->element.tag==Tag"`)
+- Adds support for element expressions of the form `element.parent==<identifier>`.
 
 ## 1.17.0 (4th January 2022)